+++ irs-ess/src/main/java/org/eclipse/tractusx/ess/discovery/EdcDiscoveryMockConfig.java
@@ -23,7 +23,6 @@
  ********************************************************************************/
 package org.eclipse.tractusx.ess.discovery;
 
-import java.util.List;
 import java.util.Map;
 
 import lombok.Data;
@@ -38,6 +37,5 @@ import org.springframework.context.annotation.Configuration;
 @ConfigurationProperties("ess.discovery")
 @Data
 public class EdcDiscoveryMockConfig {
-    private Map<String, List<String>> mockEdcAddress;
     private Map<String, SupplyChainImpacted> mockEdcResult;
 }