+++ irs-ess/src/main/java/org/eclipse/tractusx/ess/service/AspectTypeNotFoundException.java
@@ -21,38 +21,13 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.component;
-
-import com.fasterxml.jackson.annotation.JsonUnwrapped;
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Value;
-import lombok.extern.jackson.Jacksonized;
+package org.eclipse.tractusx.ess.service;
 
 /**
- * Relationship
+ * Exception thrown if ESS Validation could not find the requested aspect model in the job.
  */
-@Value
-@Jacksonized
-@Builder(toBuilder = true)
-@AllArgsConstructor
-@Schema(description = "Relationships between parent and child items.")
-public class Relationship {
-
-    private static final int GLOBAL_ASSET_ID_LENGTH = 45;
-
-    @Schema(implementation = String.class, description = "CATENA-X global asset id in the format urn:uuid:uuid4.",
-            example = "urn:uuid:6c311d29-5753-46d4-b32c-19b918ea93b0", minLength = GLOBAL_ASSET_ID_LENGTH,
-            maxLength = GLOBAL_ASSET_ID_LENGTH,
-            pattern = "^urn:uuid:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$")
-    @JsonUnwrapped
-    private GlobalAssetIdentification catenaXId;
-
-    private LinkedItem linkedItem;
-
-    private String aspectType;
-
-    private String bpn;
-
+public class AspectTypeNotFoundException extends Exception {
+    public AspectTypeNotFoundException(final String message) {
+        super(message);
+    }
 }