+++ irs-ess/src/test/java/org/eclipse/tractusx/ess/service/InvestigationJobProcessingEventListenerTest.java
@@ -33,23 +33,30 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
-import org.eclipse.tractusx.irs.edc.client.EdcSubmodelFacade;
-import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
-import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
-import org.eclipse.tractusx.ess.discovery.EdcDiscoveryFacade;
 import org.eclipse.tractusx.ess.irs.IrsFacade;
 import org.eclipse.tractusx.irs.common.JobProcessingFinishedEvent;
 import org.eclipse.tractusx.irs.component.GlobalAssetIdentification;
 import org.eclipse.tractusx.irs.component.Job;
 import org.eclipse.tractusx.irs.component.Jobs;
+import org.eclipse.tractusx.irs.component.LinkedItem;
+import org.eclipse.tractusx.irs.component.Relationship;
+import org.eclipse.tractusx.irs.component.Submodel;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.IdentifierKeyValuePair;
-import org.eclipse.tractusx.irs.component.assetadministrationshell.Reference;
 import org.eclipse.tractusx.irs.component.enums.JobState;
+import org.eclipse.tractusx.irs.data.StringMapper;
+import org.eclipse.tractusx.irs.edc.client.EdcSubmodelFacade;
+import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
+import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
+import org.eclipse.tractusx.irs.edc.client.model.notification.InvestigationNotificationContent;
+import org.eclipse.tractusx.irs.edc.client.model.notification.NotificationContent;
+import org.eclipse.tractusx.irs.registryclient.discovery.ConnectorEndpointsService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -61,32 +68,54 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class InvestigationJobProcessingEventListenerTest {
 
     private final IrsFacade irsFacade = mock(IrsFacade.class);
-    private final EdcDiscoveryFacade edcDiscoveryFacade = mock(EdcDiscoveryFacade.class);
     private final EdcSubmodelFacade edcSubmodelFacade = mock(EdcSubmodelFacade.class);
     private final BpnInvestigationJobCache bpnInvestigationJobCache = mock(BpnInvestigationJobCache.class);
     private final EssRecursiveNotificationHandler recursiveNotificationHandler = mock(
             EssRecursiveNotificationHandler.class);
     private final UUID jobId = UUID.randomUUID();
     private final UUID recursiveJobId = UUID.randomUUID();
-    private final String bpnRecursiveEdcNotification = "BPN000RECURSIVE";
+
+    private final ConnectorEndpointsService connectorEndpointsService = mock(ConnectorEndpointsService.class);
 
     private final InvestigationJobProcessingEventListener jobProcessingEventListener = new InvestigationJobProcessingEventListener(
-            irsFacade, edcDiscoveryFacade, edcSubmodelFacade, bpnInvestigationJobCache, "", "",
-            List.of(bpnRecursiveEdcNotification), recursiveNotificationHandler);
+            irsFacade, connectorEndpointsService, edcSubmodelFacade, bpnInvestigationJobCache, "", "", List.of(),
+            recursiveNotificationHandler);
 
     @Captor
-    ArgumentCaptor<EdcNotification> edcNotificationCaptor;
+    ArgumentCaptor<EdcNotification<NotificationContent>> edcNotificationCaptor;
+
+    private static AssetAdministrationShellDescriptor createShell(final String catenaXId, final String bpn) {
+        return AssetAdministrationShellDescriptor.builder()
+                                                 .globalAssetId(catenaXId)
+                                                 .specificAssetIds(List.of(IdentifierKeyValuePair.builder()
+                                                                                                 .name("manufacturerId")
+                                                                                                 .value(bpn)
+                                                                                                 .build()))
+                                                 .build();
+    }
+
+    private static Relationship createRelationship(final String lifecycle, final String bpn, final String parentId,
+            final String childId) {
+        return Relationship.builder()
+                           .aspectType(lifecycle)
+                           .bpn(bpn)
+                           .catenaXId(GlobalAssetIdentification.of(parentId))
+                           .linkedItem(
+                                   LinkedItem.builder().childCatenaXId(GlobalAssetIdentification.of(childId)).build())
+                           .build();
+    }
 
     @BeforeEach
     void mockInit() {
-        createMockForJobIdAndShell(jobId, "bpn");
+        createMockForJobIdAndShell(jobId, "bpn", List.of(createRelationship("SingleLevelBomAsPlanned", "BPN123",
+                "urn:uuid:52207a60-e541-4bea-8ec4-3172f09e6dbb", "urn:uuid:86f69643-3b90-4e34-90bf-789edcf40e7e")));
     }
 
     @Test
     void shouldSendEdcNotificationWhenJobCompleted() throws EdcClientException {
         // given
         final String edcBaseUrl = "http://edc-server-url.com";
-        when(edcDiscoveryFacade.getEdcBaseUrl(anyString())).thenReturn(Optional.of(edcBaseUrl));
+        when(connectorEndpointsService.fetchConnectorEndpoints(anyString())).thenReturn(List.of(edcBaseUrl));
         when(edcSubmodelFacade.sendNotification(anyString(), anyString(), any(EdcNotification.class))).thenReturn(
                 () -> true);
         final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(jobId.toString(),
@@ -102,9 +131,113 @@ class InvestigationJobProcessingEventListenerTest {
     }
 
     @Test
-    void shouldStopProcessingIfOneOfEdcAddressIsNotDiscovered() throws EdcClientException {
+    void shouldStopProcessingIfNoRelationshipContainsBPN() throws EdcClientException {
         // given
-        when(edcDiscoveryFacade.getEdcBaseUrl(anyString())).thenReturn(Optional.empty());
+        createMockForJobIdAndShell(jobId, "bpn",
+                List.of(createRelationship("asPlanned", null, "testParent", "testChild")));
+        final String edcBaseUrl = "http://edc-server-url.com";
+        when(connectorEndpointsService.fetchConnectorEndpoints(anyString())).thenReturn(List.of(edcBaseUrl));
+        when(edcSubmodelFacade.sendNotification(anyString(), anyString(), any(EdcNotification.class))).thenReturn(
+                () -> true);
+        final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(jobId.toString(),
+                JobState.COMPLETED.name(), "", Optional.empty());
+
+        // when
+        jobProcessingEventListener.handleJobProcessingFinishedEvent(jobProcessingFinishedEvent);
+
+        // then
+        verify(this.recursiveNotificationHandler, times(1)).handleNotification(any(), eq(SupplyChainImpacted.UNKNOWN));
+        verify(this.bpnInvestigationJobCache, times(1)).store(eq(jobId), any(BpnInvestigationJob.class));
+    }
+
+    @Test
+    void shouldHandleCaseWhenRelationshipDoesNotContainBPN() throws EdcClientException {
+        // given
+        createMockForJobIdAndShell(jobId, "bpn",
+                List.of(createRelationship("asPlanned", "BPN1", "parentId1", "childId1"),
+                        createRelationship("asPlanned", null, "parentId2", "childId2")));
+        final String edcBaseUrl = "http://edc-server-url.com";
+        when(connectorEndpointsService.fetchConnectorEndpoints(anyString())).thenReturn(List.of(edcBaseUrl));
+        when(edcSubmodelFacade.sendNotification(anyString(), anyString(), any(EdcNotification.class))).thenReturn(
+                () -> true);
+        final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(jobId.toString(),
+                JobState.COMPLETED.name(), "", Optional.empty());
+
+        // when
+        jobProcessingEventListener.handleJobProcessingFinishedEvent(jobProcessingFinishedEvent);
+
+        // then
+        verify(this.edcSubmodelFacade, times(1)).sendNotification(eq(edcBaseUrl), anyString(),
+                any(EdcNotification.class));
+        verify(this.bpnInvestigationJobCache, times(1)).store(eq(jobId), any(BpnInvestigationJob.class));
+    }
+
+    @Test
+    void shouldTriggerCorrectNotificationOnNextLevel() throws EdcClientException {
+        // given
+        createMockForJobIdAndShell(jobId, "bpn",
+                List.of(createRelationship("asPlanned", "BPN1", "parentId1", "childId1")));
+        final String edcBaseUrl = "http://edc-server-url.com";
+        when(edcSubmodelFacade.sendNotification(anyString(), anyString(), any(EdcNotification.class))).thenReturn(
+                () -> true);
+        when(connectorEndpointsService.fetchConnectorEndpoints(anyString())).thenReturn(List.of(edcBaseUrl));
+        final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(jobId.toString(),
+                JobState.COMPLETED.name(), "", Optional.empty());
+
+        // when
+        jobProcessingEventListener.handleJobProcessingFinishedEvent(jobProcessingFinishedEvent);
+
+        // then
+        verify(this.edcSubmodelFacade, times(1)).sendNotification(eq(edcBaseUrl), eq("notify-request-asset-recursive"),
+                edcNotificationCaptor.capture());
+        assertThat(edcNotificationCaptor.getValue().getHeader().getNotificationType()).isEqualTo(
+                "ess-supplier-request");
+        final InvestigationNotificationContent content = (InvestigationNotificationContent) edcNotificationCaptor.getValue()
+                                                                                                                 .getContent();
+        assertThat(content.getIncidentBPNSs()).containsAll(List.of("BPNS000000000DDD"));
+        assertThat(content.getConcernedCatenaXIds()).containsAll(List.of("childId1"));
+        verify(this.bpnInvestigationJobCache, times(1)).store(eq(jobId), any(BpnInvestigationJob.class));
+    }
+
+    @Test
+    void shouldStopProcessingIfNoEdcAddressIsDiscovered() throws EdcClientException {
+        // given
+        when(connectorEndpointsService.fetchConnectorEndpoints(anyString())).thenReturn(Collections.emptyList());
+        final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(jobId.toString(),
+                JobState.COMPLETED.name(), "", Optional.empty());
+
+        // when
+        jobProcessingEventListener.handleJobProcessingFinishedEvent(jobProcessingFinishedEvent);
+
+        // then
+        verify(this.edcSubmodelFacade, times(0)).sendNotification(anyString(), anyString(), any(EdcNotification.class));
+        verify(this.bpnInvestigationJobCache, times(1)).store(eq(jobId), any(BpnInvestigationJob.class));
+    }
+
+    @Test
+    void shouldSendCallbackIfNoMoreRelationshipsAreFound() throws EdcClientException {
+        // given
+        createMockForJobIdAndShell(jobId, "bpn", List.of());
+        when(connectorEndpointsService.fetchConnectorEndpoints(anyString())).thenReturn(Collections.emptyList());
+        final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(jobId.toString(),
+                JobState.COMPLETED.name(), "", Optional.empty());
+
+        // when
+        jobProcessingEventListener.handleJobProcessingFinishedEvent(jobProcessingFinishedEvent);
+
+        // then
+        verify(this.edcSubmodelFacade, times(0)).sendNotification(anyString(), anyString(), any(EdcNotification.class));
+        verify(this.bpnInvestigationJobCache, times(1)).store(eq(jobId), any(BpnInvestigationJob.class));
+        verify(this.recursiveNotificationHandler, times(1)).handleNotification(any(), eq(SupplyChainImpacted.NO));
+    }
+
+    @Test
+    void shouldStopProcessingIfOneOfEdcAddressesIsNotDiscovered() throws EdcClientException {
+        // given
+        createMockForJobIdAndShells(jobId, List.of("BPN123", "BPN456"));
+        final String edcBaseUrl = "http://edc-server-url.com";
+        when(connectorEndpointsService.fetchConnectorEndpoints("BPN123")).thenReturn(Collections.emptyList());
+        when(connectorEndpointsService.fetchConnectorEndpoints("BPN456")).thenReturn(List.of(edcBaseUrl));
         final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(jobId.toString(),
                 JobState.COMPLETED.name(), "", Optional.empty());
 
@@ -119,9 +252,12 @@ class InvestigationJobProcessingEventListenerTest {
     @Test
     void shouldSendEdcRecursiveNotificationWhenJobCompleted() throws EdcClientException {
         // given
-        createMockForJobIdAndShell(recursiveJobId, bpnRecursiveEdcNotification);
+        createMockForJobIdAndShell(recursiveJobId, "BPN000RECURSIVE",
+                List.of(createRelationship("SingleLevelBomAsPlanned", "BPN123",
+                        "urn:uuid:52207a60-e541-4bea-8ec4-3172f09e6dbb",
+                        "urn:uuid:86f69643-3b90-4e34-90bf-789edcf40e7e")));
         final String edcBaseUrl = "http://edc-server-url.com";
-        when(edcDiscoveryFacade.getEdcBaseUrl(anyString())).thenReturn(Optional.of(edcBaseUrl));
+        when(connectorEndpointsService.fetchConnectorEndpoints(anyString())).thenReturn(List.of(edcBaseUrl));
         when(edcSubmodelFacade.sendNotification(anyString(), anyString(), any(EdcNotification.class))).thenReturn(
                 () -> true);
         final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(
@@ -138,25 +274,98 @@ class InvestigationJobProcessingEventListenerTest {
         verify(this.bpnInvestigationJobCache, times(1)).store(eq(recursiveJobId), any(BpnInvestigationJob.class));
     }
 
-    private static AssetAdministrationShellDescriptor createShell(final String catenaXId, final String bpn) {
-        return AssetAdministrationShellDescriptor.builder()
-                                                 .globalAssetId(catenaXId)
-                                                 .specificAssetIds(List.of(IdentifierKeyValuePair.builder()
-                                                                                                 .name("manufacturerId")
-                                                                                                 .value(bpn)
-                                                                                                 .build()))
-                                                 .build();
+    @Test
+    void shouldSendEdcRecursiveNotificationWithMultipleIncidentBPNSs() throws EdcClientException {
+        // given
+        createMockForJobIdAndShell(jobId, "bpn",
+                List.of(createRelationship("asPlanned", "BPN1", "parentId1", "childId1")), List.of("BPN1", "BPN2"));
+        final String edcBaseUrl = "http://edc-server-url.com";
+        when(edcSubmodelFacade.sendNotification(anyString(), anyString(), any(EdcNotification.class))).thenReturn(
+                () -> true);
+        when(connectorEndpointsService.fetchConnectorEndpoints(anyString())).thenReturn(List.of(edcBaseUrl));
+        final JobProcessingFinishedEvent jobProcessingFinishedEvent = new JobProcessingFinishedEvent(jobId.toString(),
+                JobState.COMPLETED.name(), "", Optional.empty());
+
+        // when
+        jobProcessingEventListener.handleJobProcessingFinishedEvent(jobProcessingFinishedEvent);
+
+        // then
+        verify(this.edcSubmodelFacade, times(1)).sendNotification(eq(edcBaseUrl), eq("notify-request-asset-recursive"),
+                edcNotificationCaptor.capture());
+        final InvestigationNotificationContent content = (InvestigationNotificationContent) edcNotificationCaptor.getValue()
+                                                                                                                 .getContent();
+        assertThat(edcNotificationCaptor.getValue().getHeader().getNotificationType()).isEqualTo(
+                "ess-supplier-request");
+        assertThat(content.getIncidentBPNSs()).containsAll(List.of("BPN1", "BPN2"));
+        assertThat(content.getConcernedCatenaXIds()).containsAll(List.of("childId1"));
+        verify(this.bpnInvestigationJobCache, times(1)).store(eq(jobId), any(BpnInvestigationJob.class));
+        verify(this.edcSubmodelFacade, times(1)).sendNotification(any(), any(), any(EdcNotification.class));
     }
 
-    private void createMockForJobIdAndShell(final UUID mockedJobId, final String mockedShell) {
+    private void createMockForJobIdAndShell(final UUID mockedJobId, final String mockedShell,
+            final List<Relationship> relationships, final List<String> incindentBPNSs) {
+        final String partAsPlannedRaw = """
+                {
+                  "validityPeriod": {
+                    "validFrom": "2019-04-04T03:19:03.000Z",
+                    "validTo": "2124-12-29T10:25:12.000Z"
+                  },
+                  "catenaXId": "urn:uuid:0733946c-59c6-41ae-9570-cb43a6e4c79e",
+                  "partTypeInformation": {
+                    "manufacturerPartId": "ZX-55",
+                    "classification": "product",
+                    "nameAtManufacturer": "Vehicle Model A"
+                  }
+                }
+                """;
+        final String partSiteInformationAsPlannedRaw = """
+                {
+                  "catenaXId": "urn:uuid:0733946c-59c6-41ae-9570-cb43a6e4c79e",
+                  "sites": [
+                    {
+                      "functionValidUntil": "2025-02-08T04:30:48.000Z",
+                      "function": "production",
+                      "functionValidFrom": "2019-08-21T02:10:36.000Z",
+                      "catenaXSiteId": "BPNS000004711DMY"
+                    }
+                  ]
+                }
+                """;
+        final Submodel partAsPlanned = Submodel.from("test1", "urn:bamm:io.catenax.part_as_planned:1.0.1#PartAsPlanned",
+                StringMapper.mapFromString(partAsPlannedRaw, Map.class));
+        final Submodel partSiteInformationAsPlanned = Submodel.from("test2",
+                "urn:bamm:io.catenax.part_site_information_as_planned:1.0.0#PartSiteInformationAsPlanned",
+                StringMapper.mapFromString(partSiteInformationAsPlannedRaw, Map.class));
         final Jobs jobs = Jobs.builder()
                               .job(Job.builder()
                                       .id(mockedJobId)
                                       .globalAssetId(GlobalAssetIdentification.of("dummyGlobalAssetId"))
                                       .build())
+                              .relationships(relationships)
                               .shells(List.of(createShell(UUID.randomUUID().toString(), mockedShell)))
+                              .submodels(List.of(partAsPlanned, partSiteInformationAsPlanned))
                               .build();
-        final BpnInvestigationJob bpnInvestigationJob = BpnInvestigationJob.create(jobs, "owner", List.of("BPNS000000000DDD"));
+        final BpnInvestigationJob bpnInvestigationJob = BpnInvestigationJob.create(jobs, "owner", incindentBPNSs);
+
+        when(bpnInvestigationJobCache.findByJobId(mockedJobId)).thenReturn(Optional.of(bpnInvestigationJob));
+        when(irsFacade.getIrsJob(mockedJobId.toString())).thenReturn(jobs);
+    }
+
+    private void createMockForJobIdAndShell(final UUID mockedJobId, final String mockedShell,
+            final List<Relationship> relationships) {
+        createMockForJobIdAndShell(mockedJobId, mockedShell, relationships, List.of("BPNS000000000DDD"));
+    }
+
+    private void createMockForJobIdAndShells(final UUID mockedJobId, final List<String> bpns) {
+        final Jobs jobs = Jobs.builder()
+                              .job(Job.builder()
+                                      .id(mockedJobId)
+                                      .globalAssetId(GlobalAssetIdentification.of("dummyGlobalAssetId"))
+                                      .build())
+                              .shells(bpns.stream().map(bpn -> createShell(UUID.randomUUID().toString(), bpn)).toList())
+                              .build();
+        final BpnInvestigationJob bpnInvestigationJob = BpnInvestigationJob.create(jobs, "owner",
+                List.of("BPNS000000000DDD"));
 
         when(bpnInvestigationJobCache.findByJobId(mockedJobId)).thenReturn(Optional.of(bpnInvestigationJob));
         when(irsFacade.getIrsJob(mockedJobId.toString())).thenReturn(jobs);